LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-sans-fieldset-display-expected.txt

caption
td
td
caption
td	td
li
li
foo inline
inline
foo inline-block
inline-block
rubyrt
rubyrt

FAIL .table assert_equals: .table height expected "36px" but got "54px"
FAIL .caption assert_equals: .caption display expected "table-caption" but got "block"
FAIL .col assert_equals: .col display expected "table-column" but got "block"
FAIL .tbody assert_equals: .tbody display expected "table-row-group" but got "block"
FAIL .tr assert_equals: .tr display expected "table-row" but got "block"
FAIL .td assert_equals: .td display expected "table-cell" but got "block"
FAIL .li assert_equals: .li width expected "800px" but got "8px"
FAIL .inline assert_equals: .inline display expected "inline" but got "block"
FAIL .inline-block assert_equals: .inline-block display expected "inline-block" but got "block"
FAIL .ruby assert_equals: .ruby display expected "inline" but got "block"
FAIL .rt assert_equals: .rt width expected "29px" but got "9px"
